서버 배포를 위한 dev > main 병합 - 좋아요 기능, 로그인 및 인증 추가 (#27)
* feat: Comments 라우팅 추가

* feat: Comments 3계층 아키텍처 기초 코드 구현

* feat:게시글 에러헨들링없이 CRUD구현

* 3계층 기초 작성

* feat: Comments Create API 에러
 핸들링

* fix: Comments Create API 오류 해결 - require 오타

* feat: Comment Create API - 게시글 아이디 검증 주석 추가

* feat: Comment 목록 조회 API 1차 구현

* feat: sequelize findAll FlatObject parser 추가

* feat: Comment update API 구현, try .. catch 구문 추가

* feat: error middleware 구현

* fix: Comments Service 파일명 변경

* feat: Comment Delete API 구현

* refactor: error middleware 파일명 변경

* routes 경로 오류 수정

* feat: 게시글 조회시 댓글수 추가

* 로그인 기능 확인

* fix: app.js 미들웨어 파일 이름 오타 수정

* fix: 컨트롤러부분 수정

* fix: 꼬인거 수정

* fix:이상하게 흘러들어온거 수정

* fix: comments.routes 변수 오타 수정

* feat: 특정 댓글 조회 API Repository method 컬럼 & 조인 조건 추가

* feat: 게시글 CRUD 에러 핸들링, API명세서 맞춰서 조회기능구현

* fix:변수명 조정

* fix: 서비스내에 변수명 조정

* feat:좋아요 기능을 위한 models,migration수정+좋아요기능구현

* fix:prettierr적용안됬던거 적용

* fix:좋아요 메세지 출력값 api명세서와 달라서 수정

* fix:else if >> else로 변경

* feat: Post 모델 파일 업로드, 파일 이름 컬럼 추가

* fix: 공백 제거

* 유효성 검증 에러 미해결

* 유효성 검사 구현 중 에러 해결

* feat: CORS 설정

* fix: CORS import 추가

* fix: console.log삭제, dotenv 작성

* feat:게시글 상세 조회시 좋아요상태 좋아요 갯수 추가, 게시글 전체조회시 좋아요 갯수추가, 결과값에서 success:true 제거

* fix: 회원가입, 로그인 토큰 유효시간 수정

* fix: models파일 username으로 수정

* fix: jwt 설치, 댓글 CUD 토큰 인증 추가

* feat:게시글 CRUD에 인증 미들웨어 추가

* feat: 좋아요기능에 인증미들웨어추가

---------

+++ api/repositories/posts.repository.js
@@ -1,12 +1,24 @@
-const { Comments, Users, Posts, sequelize } = require("../../db/models");
+const { Comments, Users, Posts, Likes, sequelize } = require("../../db/models");
 const parseModelToFaltObjet = require("../helpers/parse.sequelize.helper.js");
 
 class PostsRepository {
-  createPost = async (title, content) => {
-    // const userInfo = await users.findOne({where:{userId}})
-    const createPostData = await Posts.create({ title, content });
+  createPost = async (userId, title, content) => {
+    const createPostData = await Posts.create({ userId, title, content });
 
     return createPostData;
+  };
+
+  findLikes = async () => {
+    const findLike = await Posts.findAll({
+      attributes: [
+        [sequelize.fn("COUNT", sequelize.col("Likes.postId")), "likesCount"],
+      ],
+      include: [{ model: Likes, attributes: [] }],
+      group: ["Posts.postId"],
+      order: [["createdAt", "DESC"]],
+      raw: true,
+    }).then((model) => model.map(parseModelToFaltObjet));
+    return findLike;
   };
 
   getAllPosts = async () => {
@@ -29,7 +41,9 @@ class PostsRepository {
     return allPostsData;
   };
 
-  findOnePost = async (postId) => {
+  findOnePost = async (postId, userId) => {
+    const findLikeAll = await Likes.findAll({ where: { postId } });
+    const findLike = await Likes.findOne({ where: { userId, postId } });
     const postData = await Posts.findOne({
       where: { postId },
       attributes: [
@@ -50,22 +64,27 @@ class PostsRepository {
       ],
       raw: true,
     }).then((model) => parseModelToFaltObjet(model));
+    postData.likesCount = findLikeAll.length;
+    if (!findLike) {
+      postData.likeStatus = false;
+    } else {
+      postData.likeStatus = true;
+    }
 
-    console.log(postData);
     return postData;
   };
 
-  editPost = async (postId, title, content) => {
+  editPost = async (userId, postId, title, content) => {
     const update = await Posts.update(
       { title, content },
-      { where: { postId } }
+      { where: { userId, postId } }
     );
 
     return update;
   };
 
-  deletePost = async (postId) => {
-    return await Posts.destroy({ where: { postId } });
+  deletePost = async (userId, postId) => {
+    return await Posts.destroy({ where: { userId, postId } });
   };
 }
